requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2

# Database dependencies
weaviate-client==4.4.0
neo4j==5.15.0
pymongo==4.6.0

# Agent framework
strands-agent==0.1.0  # Note: Replace with actual Strands package when available

# VAPI and OpenAI
openai==1.3.7
websockets==12.0
aiohttp==3.9.0

# MCP (Model Context Protocol)
mcp-server==0.1.0  # Note: Replace with actual MCP package

# Utilities
loguru==0.7.2
pyyaml==6.0.1
asyncio-mqtt==0.13.0
